Rename CodeMirrorComponent -> CodeEditor and ConsoleComponent -> Console
